scripts/build_and_trace_v3/requirements.txt
# 基本パッケージ
gitpython>=3.1.0
aiofiles>=23.1.0
typing-extensions>=4.7.0

# 非同期処理
asyncio>=3.4.3
aiohttp>=3.8.0

# ログ関連
colorlog>=6.7.0

# 型チェック
mypy>=1.0.0
types-aiofiles>=23.1.0
types-colorlog>=6.7.0

# テスト
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# コード品質
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# 開発ツール
ipython>=8.14.0
